identify: Restructure
Remove support for multiple importers matching the same document. This
allows the simplification of the identification process and will allow
substantial simplification of the filing and extraction. Rearragne the
code so that all console output is done in the command line interface
function.

The identify() function becomes very simple. Rewrite most of the tests
as integration doctests where testing the command line interface is
easier.