extract: Fix deduplication
The entries similarity check uses bisection to reduce the set of
entries to compare against to a narrow date interval around the date
of the newly extracted entry. This works only of the existing entries
are sorted by date. Sort the existing entries array.

Update the find_similar_entries() docstring to state this and change
parameter names to be more consisten with what is used in other parts
of the extraction code.