extract: Implement entries deduplication within the same extract run
After extracting entries from all the documents, sort the entries
grouped by source document on the base of the dates and document
accout. Then iteratively run the deduplication on each group,
appending the new entries to the list of existing entries.

Sorting the entries groups allows for reproducible and predictable
output without relying on the documents filenames.

This changes the interface of the extraction hooks: now the hooks
receive a (filename, entries, account, importer) tuple.  The importer
is not used yet, but it will be used to allow importers to customize
deduplication.